Added the base of adding online database

+++ app/src/main/java/ir/hamedan/budgetmanagement/data/repository/SavingGoalRepository.kt
@@ -4,11 +4,27 @@ import ir.hamedan.budgetmanagement.data.local.models.SavingGoalEntity
 import kotlinx.coroutines.flow.Flow
 
 interface SavingGoalRepository {
+
     fun getAllGoals(): Flow<List<SavingGoalEntity>>
+
     suspend fun insertGoal(goal: SavingGoalEntity)
+
     suspend fun updateGoal(goal: SavingGoalEntity)
+
     suspend fun deleteGoal(goal: SavingGoalEntity)
-    suspend fun depositToGoal(goalId: String, amount: Double)
-    suspend fun withdrawFromGoal(goalId: String, amount: Double)
-    suspend fun updateLastAutoDepositTimestamp(goalId: String, timestamp: Long)
+
+    suspend fun depositToGoal(
+        goalId: String,
+        amount: Long
+    )
+
+    suspend fun withdrawFromGoal(
+        goalId: String,
+        amount: Long
+    )
+
+    suspend fun updateLastAutoDepositTimestamp(
+        goalId: String,
+        timestamp: Long
+    )
 }
